% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAlerts.R
\name{getAlerts}
\alias{getAlerts}
\title{Get active National Weather Service alerts}
\usage{
getAlerts(includeStates = NULL, excludeStates = NULL, spatial = TRUE)
}
\arguments{
\item{includeStates}{2-letter state abbreviations for the states for which
alerts should be retrieved. Default is \code{NULL} indicating all 50 states
and the District of Columbia are included.}

\item{excludeStates}{2-letter state abbreviations for the states to exclude
from the output. A common choice is \code{c("AK", "HI")}. Default is
\code{NULL} indicating no states are excluded.}

\item{spatial}{logical indicating whether to return alert area geometries. See value.}
}
\value{
If package \pkg{weatherAlertAreas} is installed, then a
  \code{\link[sp:SpatialPolygonsDataFrame-class]{SpatialPolygonsDataFrame}}
  containing alert information and corresponding alert area geometries. If
  package \pkg{weatherAlertAreas} is not installed, then a data frame
  containing alert information.
}
\description{
Retrieve information about current United States National Weather Service
(\acronym{NWS}) watches, warnings, and advisories.
}
\details{
Alerts are currently limited to the 50 United States and the
  District of Columbia.
}
\seealso{
Package \pkg{weatherAlertAreas}
}

